requirements.txt
pytorch==0.4.0
numpy==1.13.1
tqdm==4.23.4
zhon==1.1.5
pickle